requirements.txt
# Core FastAPI dependencies
fastapi==0.124.4
uvicorn[standard]==0.38.0

# Database
SQLAlchemy==2.0.45
asyncpg==0.31.0

# Authentication & Security
python-jose[cryptography]==3.5.0
passlib[argon2]==1.7.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0

# Data validation
pydantic==2.12.5
pydantic-settings==2.12.0

# Database migrations
alembic==1.16.0

# Environment variables
python-dotenv==1.2.1
